Move shared components to the shared package
