+++ web/packages/shared/components/UnifiedResources/ResourceTab.tsx
@@ -18,7 +18,10 @@ import React from 'react';
 import styled from 'styled-components';
 import { Box, Text } from 'design';
 
-import { HoverTooltip, PINNING_NOT_SUPPORTED_MESSAGE } from './Resources';
+import {
+  HoverTooltip,
+  PINNING_NOT_SUPPORTED_MESSAGE,
+} from './UnifiedResources';
 
 export const ResourceTab = ({
   title,
