+++ web/packages/teleport/src/components/hooks/index.ts
@@ -16,4 +16,3 @@
 
 export { useUrlFiltering } from './useUrlFiltering';
 export { useServerSidePagination } from './useServersidePagination';
-export { useInfiniteScroll } from './useInfiniteScroll';
